Added explorer events.

+++ src/main/java/net/jamsimulator/jams/gui/action/defaults/explorerelement/mips/filestoassemble/MipsFilesToAssembleActionRemove.java
@@ -36,6 +36,7 @@ import net.jamsimulator.jams.gui.mips.sidebar.FilesToAssembleDisplay;
 import net.jamsimulator.jams.gui.mips.sidebar.FilesToAssembleDisplayElement;
 import net.jamsimulator.jams.gui.project.ProjectTab;
 import net.jamsimulator.jams.language.Messages;
+import net.jamsimulator.jams.project.Project;
 import net.jamsimulator.jams.project.mips.MipsFilesToAssemble;
 import net.jamsimulator.jams.project.mips.MipsProject;
 
@@ -63,8 +64,9 @@ public class MipsFilesToAssembleActionRemove extends ContextAction {
 
 		ProjectTab tab = JamsApplication.getProjectsTabPane().getFocusedProject().orElse(null);
 		if (tab == null) return;
-		MipsProject project = tab.getProject();
-		MipsFilesToAssemble files = project.getData().getFilesToAssemble();
+		Project project = tab.getProject();
+		if (!(project instanceof MipsProject)) return;
+		MipsFilesToAssemble files = ((MipsProject) project).getData().getFilesToAssemble();
 
 		for (ExplorerElement element : elements) {
 			files.removeFile(((FilesToAssembleDisplayElement) element).getFile());
